+++ jni/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## Next released
+
+### Added
+
+* `withLocalRef`: a helper to delete local references.
+
 ## [0.5.0] - 2017-12-02
 
 ### Added